Screen selection works again

+++ lib/miku/midi/midi.hpp
@@ -9,8 +9,8 @@ namespace miku::midi {
 
     const std::string notes[] = {"C", "C#", "D", "D#", "E", "F", "F#", "G", "G#", "A", "A#", "B"};
 
-    std::string GetNameFromNoteNumber(int note) {
-        return notes[note % 12] + std::to_string(note / 12 - 1);
+    std::string GetNameFromNoteNumber(uint8_t note) {
+        return notes[note % 12] + std::to_string(note / (12 - 1));
     }
 
     std::string GetMidiTypeAsString(daisy::MidiEvent& msg, char* str)
